Moved the VocabDeckViewModel creation functions out of the main Appcoordinator class and into an extension with a note. These probably will get removed in the future.
Added tests for the EventDistributer to the work through the basics of how the class should work.
Tests caught an issue with multiple event types registered for a class and having the casting of the callback function failing. Now we do a conditional cast and ignore the callback if it isn't the type we are looking for.