WIP to get the parsed json data to be represented in the UI.
Switched DeckCollection back to a protocol from a struct. The DeckCollection concrete type, DeckCollectionViewModel  keeps a private reference to the VocabulatorDeckCollection model object but the VocabDeckPickerViewModel doesn't need to know anything about that. Using a protocol hides that extra information.
Switched the DeckViewModel to be DeckCellViewModel to keep naming consistent and added an EmptyDeckCollection concrete type that represents an empty collection.
DeckCollections now keep an array of WordDeck types instead of direct refs to DeckCellViewModel and VocabDeckPickerViewModel only references WordDeck instead of the viewModel type. Will eventually move the concrete DeckCellViewModel into a separate file.

Implemented the VocabulatorDataModel to wrap a private singleton DataProvider type that knows how to load json from urls. VocabulatorDataModel provides a few static methods for accessing the data and currently all of this is on the main thread. Will need to look at pushing parsing and loading off onto background queues at some point. Currently the VocabulatorDataModel only knows how to load the demo data, but there is a todo() for extending functionality for loading collection json documents in the Documents folder (once that functionality is built).

Updated AppCoordinator to use the VocabulatorDataModel and added a initialViewModel() func that creates the initial view model for the VocabDeckPickerCollectionViewController.
Added the DeckCollecitonViewModel to the AppCoordinator.swift file. All the properties are getters that reach into a private backing VocabulatorDeckCollection. Also using a lazily mapped collection for the decks array. It coverts to a WordDeck type from the VocabulatorDeck type on demand.
Updated the selection handling to find the selected deck and pass it to the WordViewerCoordinator when a deck is selected.

Updated WordViewerCoordinator to take a VocabulartorDeck as an argument to the default init. Nothing else is being done with the data currently.

Fixed a bug in the DemoWords.json data that was causing the JSON parser to fail and moved the bottom constraint of the VocabDeckColelctionCell's title label to be a little lower.