Some rethinking of how the WordProvider internals work.
Switched to using an integrator on the deck.words array and having the provider just keep track of the current index, the iterator does the rest.
Also added a reset() function that can reset the provider back to the beginning again

Added tests to drive out the reset functionality and reran all tests to make sure behavior didn't change after the iterator switch.