Work to integrate the WordProvider and the WordViewerCoordinator together to drive showing the next word.
Added a tap gesture recognizer to the WordViewerViewController that fires the WordEvents.next() action. This event is picked up by the WordViewerCoordinator and it uses a couple of new internal helper functions to create the view model and the next view controller for the next word.
Exposed the currentWordIndex getter on the WordProvider so that the (n of n) count string can rely on the same data as the word we are using.

Still have a bug where we crash if after the last word the user taps again because we end up trying to stuff a nil view model into the next WordViewerViewController.